Organize the file structure
